Added `textInput` element with support for translating attributes (#2)
It seems that customized built-in elements are not well supported by browsers,
Safari won't implement it so we'll have to create minimal proxy elements. This
may provide us with the opportunity to simplify the DOM node API for our set of
custom elements. The render logic may also be easily shareable.