Scrapping global configuration by looking for the nearest <intl-context> (#2)
To enable the assembly of dynamically loadable page fragments in the future the
individual parts need to be able to set up their own `intl` contexts. Using a
simple <intl-context> to hang off the formerly global `intl` instance allows
easy discovery by <intl-elements> inside the sub-tree.

We still need to find a way to properly configure the context, preferably
with declarative attributes for easy Elm interop.